requirements.txt
word4univer~=0.1.0
setuptools~=60.2.0
python-telegram-bot==20.0a4
loguru~=0.6.0